% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_assignment.R
\name{assert_assignment_config}
\alias{assert_assignment_config}
\title{Assert a \code{assignment_config} object.}
\usage{
assert_assignment_config(assignment)
}
\arguments{
\item{assignment}{object to test.}
}
\value{
a checked \code{assignment_config} object.
}
\description{
Check assignment yml file that it is a correct assignment file.
}
\keyword{internal}
