% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_assignment.R
\name{assignment_config}
\alias{assignment_config}
\title{Constructor for \code{assignment_config} object}
\usage{
assignment_config(x)
}
\arguments{
\item{x}{a list to convert to a \code{assignment_config} object.}
}
\description{
Constructor for \code{assignment_config} object.
}
\keyword{internal}
