% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_assignment.R
\name{assignment_paths_and_files}
\alias{assignment_paths_and_files}
\title{Get assignment paths and files}
\usage{
assignment_paths_and_files(assignment = NULL)
}
\arguments{
\item{assignment}{a \code{assignment_config} object.}
}
\description{
Get assignment paths and files
}
\keyword{internal}
