% Generated by roxygen2 (4.0.2): do not edit by hand
\name{assignment_yml_ok}
\alias{assignment_yml_ok}
\title{Check the yml file to be a correct assignment yml.}
\usage{
assignment_yml_ok(path = NULL)
}
\value{
boolean
}
\description{
Check the yml file to be a correct assignment yml.
}
\details{
Check the yml file that it is correct and can be used. Otherwise warn.
}

