% Generated by roxygen2 (4.0.2): do not edit by hand
\name{cached_tasks}
\alias{cached_tasks}
\title{Cached tasks}
\usage{
cached_tasks()
}
\value{
character vector with cached assignments.
}
\description{
Cached tasks
}
\details{
Checks which assignments that are cached (ie already downloaded to temp dir).
}

