% Generated by roxygen2 (4.0.2): do not edit by hand
\name{check_assignment_file}
\alias{check_assignment_file}
\title{Check assignment yml file that it is a correct assignment file}
\usage{
check_assignment_file(assignment)
}
\arguments{
\item{assignment}{list to test.}
}
\value{
boolean
}
\description{
Check assignment yml file that it is a correct assignment file
}

