% Generated by roxygen2 (4.0.2): do not edit by hand
\name{check_assignment_yml}
\alias{check_assignment_yml}
\title{Check yml file to be correct.}
\usage{
check_assignment_yml()
}
\value{
boolean
}
\description{
Check yml file to be correct.
}
\details{
Check the yml file that it is correct and can be used. Otherwise warn.
}

