% Generated by roxygen2 (4.0.2): do not edit by hand
\name{check_cached_assignments}
\alias{check_cached_assignments}
\title{Check for cached assignments.}
\usage{
check_cached_assignments()
}
\value{
character vector with cached assignments.
}
\description{
Check for cached assignments.
}
\details{
Checks which assignments that are cached (ie downloaded to temp dir).
}

