% Generated by roxygen2 (4.0.1): do not edit by hand
\name{check_config}
\alias{check_config}
\title{Check the config file.}
\usage{
check_config(config_list)
}
\arguments{
\item{test_suite}{Path to test suite.}
}
\description{
Check the config file.
}
\details{
Check that the test suites in the config file is possible to use
}
\examples{
get_test_suite_type("https://api.github.com/repos/MansMeg/KursRprgm/contents/Labs/Test/")
}

