% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertion_funcions.R
\name{check_existance_tasks}
\alias{check_existance_tasks}
\title{Check that tasks exist in assignment.}
\usage{
check_existance_tasks(tasks, path = NULL)
}
\arguments{
\item{tasks}{The \code{task} vector from \code{mark_my_assignment}.}

\item{path}{Path to assignment file. Passed to \code{read_assignment_yml}.}
}
\value{
A warning message or nothing.
}
\description{
Checks that tasks that are inputted in mark_my_assignment exists. If not, a warning is produced.
}
\keyword{internal}

