% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_assignment.R
\name{construct_assignment_description}
\alias{construct_assignment_description}
\title{Construct assignment description}
\usage{
construct_assignment_description(assignment = NULL)
}
\arguments{
\item{assignment}{an assignment to create description for.}
}
\description{
Construct assignment description
}
\keyword{internal}
