% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github.R
\name{create_github_download_url}
\alias{create_github_download_url}
\title{Create a github_download_url from a path_github object.}
\usage{
create_github_download_url(path)
}
\arguments{
\item{path}{a \code{path_github} object.}
}
\description{
Create a github_download_url from a path_github object.
}
\keyword{internal}
