% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mark_my_assignment.R
\name{data_mma}
\alias{data_mma}
\title{Loads a data set into a specific environment}
\usage{
data_mma(..., env)
}
\arguments{
\item{...}{Arguments to pass on to data().}

\item{env}{Environment to load data into.}
}
\value{
One or more data sets are loaded into the specified environment
}
\description{
Loads a data set into a specific environment
}

