% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mark_my_assignment.R
\name{delete_circular_calls}
\alias{delete_circular_calls}
\title{Checks and deletes circular calls}
\usage{
delete_circular_calls(mark_file)
}
\arguments{
\item{mark_file}{File to check}
}
\value{
Character vector of the possibly changed mark file
}
\description{
Checks and deletes circular calls
}
\keyword{internal}

