% Generated by roxygen2 (4.0.2): do not edit by hand
\name{directories}
\alias{directories}
\alias{mark_my_assignment_dir}
\alias{mark_my_base_dir}
\alias{mark_my_test_dir}
\title{Functions to create directories}
\usage{
mark_my_base_dir()

mark_my_assignment_dir(no = 1)

mark_my_test_dir(...)
}
\description{
Functions to create directories
}

