% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_assignment.R
\name{download_assignment}
\alias{download_assignment}
\title{Download assignment and store in temporary folder}
\usage{
download_assignment(assignment, path)
}
\description{
Downloads the test files for the current assignment and save them to 
temp directory.
}
\keyword{internal}
