% Generated by roxygen2 (4.0.2): do not edit by hand
\name{expect_arguments}
\alias{expect_arguments}
\title{Expect object}
\usage{
expect_arguments(object, ..., info = NULL, label = NULL)
}
\arguments{
\item{object}{Object that is expected to exist.}
}
\description{
Expect object
}
\details{
Test that an object with a given name exist in the environment.
}

