% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expectations.R
\name{expect_attached_package}
\alias{expect_attached_package}
\title{Expect that a given package is used}
\usage{
expect_attached_package(object, info = NULL)
}
\arguments{
\item{object}{Package to check for.}

\item{info}{Extra information to be included in the message (useful when writing tests in loops).}
}
\description{
Tests that the following packages is used.
}
\keyword{internal}

