% Generated by roxygen2 (4.0.2): do not edit by hand
\name{expect_function_arguments}
\alias{expect_function_arguments}
\title{Expect function arguments}
\usage{
expect_function_arguments(object, expected, ..., info = NULL, label = NULL,
  expected.label = NULL)
}
\arguments{
\item{object}{Object that is expected to exist.}
}
\description{
Expect function arguments
}
\details{
Test that an object with a given name exist in the environment.
}

