% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expectations.R
\name{expect_function_code}
\alias{expect_function_code}
\title{Expect function contain code}
\usage{
expect_function_code(object, expected, info = NULL, label = NULL,
  expected.label = NULL)
}
\arguments{
\item{object}{Function to check for mandatory code}

\item{expected}{Expected arguments in function.}

\item{info}{Extra information to be included in the message (useful when writing tests in loops).}

\item{label}{For full form, label of expected object used in error messages. 
Useful to override default (deparsed expected expression) when doing 
tests in a loop. For short cut form, object label. When NULL, computed from 
deparsed object.}

\item{expected.label}{Equivalent of \code{label} for shortcut form.}
}
\description{
Test that a given code code exists in function
}
\keyword{internal}

