% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expectations.R
\name{expect_no_attached_forbidden_package}
\alias{expect_no_attached_forbidden_package}
\title{Expect that a forbidden package is not used/attached}
\usage{
expect_no_attached_forbidden_package(pkg)
}
\arguments{
\item{pkg}{Package to check for.}
}
\description{
Tests that the following packages is not used.
}
\keyword{internal}
