% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expectations.R
\name{expect_no_forbidden_function_code}
\alias{expect_no_forbidden_function_code}
\title{Expect no forbidden function code}
\usage{
expect_no_forbidden_function_code(object, forbidden)
}
\arguments{
\item{object}{Function to check for mandatory code}

\item{forbidden}{Code string that are forbidden to use.}
}
\description{
Test that a given code string does not exists in function.
}
\keyword{internal}
