% Generated by roxygen2 (4.0.2): do not edit by hand
\name{expect_not_use_package}
\alias{expect_not_use_package}
\title{Expect that packages are not used}
\usage{
expect_not_use_package()
}
\arguments{
\item{pkg}{Package to check for.}
}
\description{
Expect that packages are not used
}
\details{
Tests that the following packages are not used.
}

