% Generated by roxygen2 (4.0.2): do not edit by hand
\name{expect_package_not_used}
\alias{expect_package_not_used}
\title{Expect that packages are not used}
\usage{
expect_package_not_used(object, ..., info = NULL, label = NULL)
}
\arguments{
\item{pkg}{Package to check for.}
}
\description{
Expect that packages are not used
}
\details{
Tests that the following packages are not used.
}

