% Generated by roxygen2 (4.0.2): do not edit by hand
\name{expect_package_used}
\alias{expect_package_used}
\title{Expect that a given package is used}
\usage{
expect_package_used(object, info = NULL, label = NULL)
}
\arguments{
\item{object}{Package to check for.}

\item{info}{Extra information to be included in the message (useful when writing tests in loops).}

\item{label}{For full form, label of expected object used in error messages.
Useful to override default (deparsed expected expression) when doing
tests in a loop. For short cut form, object label. When NULL, computed from
deparsed object.}
}
\description{
Expect that a given package is used
}
\details{
Tests that the following packages is used.
}

