% Generated by roxygen2 (4.0.2): do not edit by hand
\name{expect_tidy_code}
\alias{expect_tidy_code}
\title{Expect tidy format (to be constructed)}
\usage{
expect_tidy_code()
}
\description{
Expect tidy format (to be constructed)
}
\details{
Test that the format used in a function is tidy (see formatR)
}

