% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mark_my_assignment.R
\name{forbidden_functions}
\alias{forbidden_functions}
\title{Forbidden functions for \code{mark_my_file()}}
\usage{
forbidden_functions()
}
\description{
Funcations that are removed prior to running \code{mark_my_file()}. 
The purpose is to avoid reinstalling packages, doing system calls or similar when checking a file.
}
\keyword{internal}
