% Generated by roxygen2 (4.0.2): do not edit by hand
\name{function_code}
\alias{function_code}
\title{Function contain code test}
\usage{
function_code(expected, label = NULL)
}
\arguments{
\item{expected}{Pattern to test for in function code.}

\item{label}{Expectation label used by \code{expect_function_code()}}
}
\description{
Function contain code test
}
\details{
Test if function code contains a given text string.
}

