% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_assignment.R
\name{get_assignment_full_subpath}
\alias{get_assignment_full_subpath}
\title{Get the full path from assignment paths}
\usage{
get_assignment_full_subpath(sub_path, path)
}
\arguments{
\item{sub_path}{the url to compute the full path for.}

\item{path}{the assignment \code{path_type}.}
}
\description{
The individual paths of assignmentfiles etc can be specified both in full and relative to the assignment file.
This function computes the full path.
}
\keyword{internal}
