% Generated by roxygen2 (4.0.2): do not edit by hand
\name{get_file}
\alias{get_file}
\title{Get the file from the path}
\usage{
get_file(path, dest, ...)
}
\arguments{
\item{path}{Path object}

\item{dest}{Destination for the file}

\item{...}{Further arguments to send to \code{httr::GET()}.}
}
\description{
Get the file from the path
}
\details{
Get/download the file from the path.
}

