% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github.R
\name{get_github_path_info}
\alias{get_github_path_info}
\title{Get github information from a github path}
\usage{
get_github_path_info(path)
}
\arguments{
\item{path}{a path to check}
}
\value{
a list
}
\description{
Get github information from a github path
}
\keyword{internal}
