% Generated by roxygen2 (4.0.1): do not edit by hand
\name{get_test_suite_type}
\alias{get_test_suite_type}
\title{Get type of source for test suite.}
\usage{
get_test_suite_type(test_suite)
}
\arguments{
\item{test_suite}{Path to test suite.}
}
\description{
Get type of source for test suite.
}
\details{
Computes the type of test suite source.
}
\examples{
get_test_suite_type("https://api.github.com/repos/MansMeg/KursRprgm/contents/Labs/Test/")
}

