% Generated by roxygen2 (4.0.2): do not edit by hand
\name{get_test_that_files}
\alias{get_test_that_files}
\title{Get test files}
\usage{
get_test_that_files(task = "all")
}
\arguments{
\item{task}{Which task should be downloaded. Default is "all".}
}
\description{
Get test files
}
\details{
Downloads the test files for the current assignment and save them to
temp directory.
}

