% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expectations.R
\name{has_function_arguments}
\alias{has_function_arguments}
\title{Function has argument test}
\usage{
has_function_arguments(expected, label = NULL)
}
\arguments{
\item{expected}{Arguments as text vector to test for.}

\item{label}{Expectation label used by \code{expect_function_arguments()}}
}
\description{
Test if a function has the given arguments
}
\keyword{internal}

