% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github.R
\name{is_github_path}
\alias{is_github_path}
\title{Check if a path is from github}
\usage{
is_github_path(path)
}
\arguments{
\item{path}{a path to check}
}
\value{
a \code{bool} that is named \code{raw}, \code{http} or \code{api} if a github path.
}
\description{
Check if a path is from github
}
\keyword{internal}
