% Generated by roxygen2 (4.0.2): do not edit by hand
\name{is_self_contained}
\alias{is_self_contained}
\title{Function is self contained test}
\usage{
is_self_contained(expected)
}
\arguments{
\item{expected}{Function to test if it is self contained.}
}
\description{
Function is self contained test
}
\details{
Tests if a function is self contained (no global variables)
}

