% Generated by roxygen2 (4.0.2): do not edit by hand
\name{load_assignment_info}
\alias{load_assignment_info}
\title{Load assignment information}
\usage{
load_assignment_info()
}
\value{
assignment object
}
\description{
Load assignment information
}
\details{
Check if there is an assignmentfile and load it.
}

