% Generated by roxygen2 (4.0.2): do not edit by hand
\name{path_type}
\alias{path_type}
\title{Get the path type.}
\usage{
path_type(path)
}
\arguments{
\item{path}{Character element of url or local search path.}
}
\value{
path type as character element c("path_local", "path_http", "path_error")
}
\description{
Get the path type.
}
\details{
Check the path type.
}

