% Generated by roxygen2 (4.0.2): do not edit by hand
\name{read_assignment_yml}
\alias{read_assignment_yml}
\title{Load assignment information}
\usage{
read_assignment_yml()
}
\value{
assignment object
}
\description{
Load assignment information
}
\details{
Check if there exist an assignmentfile and then load it.
}

