% Generated by roxygen2 (4.0.2): do not edit by hand
\name{run_test_suite}
\alias{run_test_suite}
\title{Run test suite.}
\usage{
run_test_suite()
}
\arguments{
\item{task}{Which task should be tested}

\item{file}{Run tests on a R-file. Default is NULL means global environment.}
}
\value{
Character vector with cached assignments.
}
\description{
Run test suite.
}
\details{
Runs test on the tasks. Always run mandatory tests.
}

