% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_assignment.R
\name{set_assignment}
\alias{set_assignment}
\alias{remove_assignment}
\title{Set assignment to mark}
\usage{
set_assignment(path, auth = NULL)

remove_assignment()
}
\arguments{
\item{path}{Path to the yml file}

\item{auth}{Authorization object generated by \code{httr::authenticate()}. See \code{authenticate} and \url{https://github.com/r-lib/httr/blob/master/demo/oauth2-github.r} on details on how to use \code{authenticate()}.}
}
\description{
Sets the assignment to mark and downloads necessary files.
}
\examples{
assignment_path <- path <- 
  paste0(system.file(package = "markmyassignment"), "/extdata/example_assignment01.yml")
set_assignment(assignment_path)

}
