% Generated by roxygen2 (4.0.1): do not edit by hand
\name{set_test_suite}
\alias{set_test_suite}
\title{Set source for testsuite}
\usage{
set_test_suite(test_suite)
}
\arguments{
\item{test_suite}{Test source}
}
\description{
Set source for testsuite
}
\details{
Currently the following sources can be used for storing testsuites.
- Github API adress to folder
- Local directory

Multiple sources can be set
}
\examples{
set_test_suite(test_suite = "https://api.github.com/repos/MansMeg/KursRprgm/contents/Labs/Test/")
}

