% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_assignment.R
\name{show_tasks}
\alias{show_tasks}
\title{Get the name of the tasks in the assignment.}
\usage{
show_tasks()
}
\description{
Get the name of the tasks in the assignment.
}
\examples{
\donttest{
assignment_path <- 
  paste0(system.file(package = "markmyassignment"), "/extdata/example_assignment.yml.R")
set_assignment(assignment_url)
show_tasks()
}

}

