% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mark_my_assignment.R
\name{stop_if_circular_calls}
\alias{stop_if_circular_calls}
\title{Checks and stop if there are circular calls}
\usage{
stop_if_circular_calls(mark_file)
}
\arguments{
\item{mark_file}{File to check}
}
\description{
Checks and stop if there are circular calls
}

