% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporter.R
\docType{class}
\name{StudentReporter-class}
\alias{StudentReporter}
\alias{StudentReporter-class}
\title{Test reporter: Simplified summary of failures and errors.}
\arguments{
\item{...}{Arguments used to initialise class}
}
\description{
This reporter is used to report errors and failures to students. Only
the testname and test information is returned. Otherwise similar to
SummaryReporter.
}
\keyword{debugging}
\keyword{internal}

