% Generated by roxygen2 (4.0.2): do not edit by hand
\name{temp_folder_check_create}
\alias{temp_folder_check_create}
\title{Check and create folder if missing.}
\usage{
temp_folder_check_create()
}
\description{
Check and create folder if missing.
}
\details{
Checks if markmyassignment folder exist in R temp directory.
  If not, the folder is created.
}

