% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mark_my_assignment.R
\name{translate_tasks_name_to_task_files}
\alias{translate_tasks_name_to_task_files}
\title{Get task file name from task names}
\usage{
translate_tasks_name_to_task_files(tasks)
}
\arguments{
\item{tasks}{task names in assignment}
}
\description{
Get task file name from task names
}
\keyword{internal}
