% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_assignment.R
\name{write_assignment_yml}
\alias{write_assignment_yml}
\title{Read assignment information}
\usage{
write_assignment_yml(assignment, yml_path)
}
\arguments{
\item{assignment}{a \code{assignment_config} object to write}

\item{yml_path}{a \code{path} to write yml file to}
}
\description{
Check if there exist an assignmentfile and then load it.
}
\keyword{internal}
