% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertion_funcions.R
\name{assert_function_arguments_in_API}
\alias{assert_function_arguments_in_API}
\title{Checks the input arguments for mark_my_assignment and mark_my_file}
\usage{
assert_function_arguments_in_API(tasks, mark_file, lab_file = NULL,
  force_get_tests, quiet, reporter)
}
\arguments{
\item{tasks}{Arguments from \code{mark_my_assignment} and \code{mark_my_file}.}

\item{mark_file}{Arguments from \code{mark_my_assignment} and \code{mark_my_file}.}

\item{lab_file}{Arguments from \code{mark_my_assignment} and \code{mark_my_file}.}

\item{force_get_tests}{Arguments from \code{mark_my_assignment} and \code{mark_my_file}.}

\item{quiet}{Arguments from \code{mark_my_assignment} and \code{mark_my_file}.}

\item{reporter}{Arguments from \code{mark_my_assignment} and \code{mark_my_file}.}
}
\value{
If all inputs are OK, nothing. Otherwise the functions stop.
}
\description{
Checks that the input arguments are of the correct type.
}
\keyword{internal}

