% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_assignment.R
\name{assignment_yml_ok}
\alias{assignment_yml_ok}
\title{Check the yml file to be a correct assignment yml.}
\usage{
assignment_yml_ok(path = NULL)
}
\arguments{
\item{path}{\code{path} object from \code{\link{path_type}}}
}
\value{
boolean
}
\description{
Check the yml file that it is correct and can be used. Otherwise warn.
}

