% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mark_my_assignment.R
\name{cached_tasks}
\alias{cached_tasks}
\title{Cached tasks}
\usage{
cached_tasks()
}
\value{
character vector with cached assignments.
}
\description{
Checks which assignments that are cached (ie already downloaded to temp dir).
}
\keyword{internal}

