% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_assignment.R
\name{check_installed_packages}
\alias{check_installed_packages}
\title{Check whether required packages are installed and loaded.}
\usage{
check_installed_packages(packages)
}
\arguments{
\item{packages}{Packages to check}
}
\description{
Checks if the packages listed in assignment file are loaded and installed.
  If not, a warning message is printed.
}
\keyword{internal}

