% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_assignment.R
\name{directories}
\alias{directories}
\alias{mark_my_base_dir}
\alias{mark_my_assignment_dir}
\alias{mark_my_tasks_dir}
\alias{mark_my_run_code_dir}
\title{Functions to create directories}
\usage{
mark_my_base_dir()

mark_my_assignment_dir(no = 1)

mark_my_tasks_dir(...)

mark_my_run_code_dir(...)
}
\arguments{
\item{no}{assignment number}

\item{...}{to send to \code{\link{mark_my_assignment_dir}}}
}
\description{
Functions to create directories
}
\keyword{internal}
