% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expectations.R
\name{expect_attached_package}
\alias{expect_attached_package}
\title{Expect that a given package is used}
\usage{
expect_attached_package(pkg, info = NULL)
}
\arguments{
\item{pkg}{Package to check for.}

\item{info}{Deprecated.}
}
\description{
Tests that the following packages is used.
}
\keyword{internal}
