% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expectations.R
\name{expect_function_arguments}
\alias{expect_function_arguments}
\title{Expect function arguments}
\usage{
expect_function_arguments(object, expected, info = NULL, label = NULL,
  expected.label = NULL)
}
\arguments{
\item{object}{Function to check the arguments of.}

\item{expected}{Expected arguments in function.}

\item{info}{Extra information to be included in the message (useful when writing tests in loops).}

\item{label}{For full form, label of expected object used in error messages. 
Useful to override default (deparsed expected expression) when doing 
tests in a loop. For short cut form, object label. When NULL, computed from 
deparsed object.}

\item{expected.label}{Equivalent of \code{label} for shortcut form.}
}
\description{
Test that an object with a given name exist in the environment.
}

