% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expectations.R
\name{expect_function_arguments}
\alias{expect_function_arguments}
\title{Expect function arguments}
\usage{
expect_function_arguments(object, expected, info = NULL, label = NULL,
  expected.label = NULL)
}
\arguments{
\item{object}{Function to check the arguments of.}

\item{expected}{Expected arguments in function.}

\item{info}{Deprecated.}

\item{label}{Additional information.}

\item{expected.label}{Deprecated.}
}
\description{
Test that an function object has a function with given arguments.
}
\keyword{internal}
