% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expectations.R
\name{expect_function_code}
\alias{expect_function_code}
\title{Expect function contain code}
\usage{
expect_function_code(object, expected, info = NULL, label = NULL,
  expected.label = NULL)
}
\arguments{
\item{object}{Function to check for mandatory code}

\item{expected}{Expected arguments in function.}

\item{info}{Deprecated.}

\item{label}{Additional information.}

\item{expected.label}{Deprecated.}
}
\description{
Test that a given code string exists in function
}
\keyword{internal}
