% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expectations.R
\name{expect_package}
\alias{expect_package}
\title{Depricated function: expect_package}
\usage{
expect_package(object, info = NULL, label = NULL)
}
\description{
Function has been depricated and will be removed. Please use \code{\link{expect_attached_package}} instead.
}
\keyword{internal}
