% Generated by roxygen2 (4.0.2): do not edit by hand
\name{expect_self_contained}
\alias{expect_self_contained}
\title{Expect that the tested function is self-contained.}
\usage{
expect_self_contained(object, info = NULL, label = NULL)
}
\arguments{
\item{object}{Function to test if it is self-contained.}

\item{info}{Extra information to be included in the message (useful when writing tests in loops).}

\item{label}{For full form, label of expected object used in error messages.
Useful to override default (deparsed expected expression) when doing
tests in a loop. For short cut form, object label. When NULL, computed from
deparsed object.}
}
\description{
Expect that the tested function is self-contained.
}
\details{
Tests if a fuction is self-contained (i.e. do not use any global variables).
}

