% Generated by roxygen2 (4.0.2): do not edit by hand
\name{expect_self_contained}
\alias{expect_self_contained}
\title{Expect that the tested function is self-contained.}
\usage{
expect_self_contained(object, ..., info = NULL, label = NULL)
}
\arguments{
\item{object}{Function to test if it is self-contained.}
}
\description{
Expect that the tested function is self-contained.
}
\details{
Tests if a fuction is self-contained (i.e. do not use any global variables).
}

