% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expectations.R
\name{expect_self_contained}
\alias{expect_self_contained}
\title{Depricated function: expect_self_contained}
\usage{
expect_self_contained(object, info = NULL, label = NULL)
}
\description{
Function has been depricated and will be removed. Please use \code{\link{expect_function_self_contained}} instead.
}
\keyword{internal}

