% Generated by roxygen2 (4.0.2): do not edit by hand
\name{expect_tidy_code}
\alias{expect_tidy_code}
\title{Expect tidy format}
\usage{
expect_tidy_code()
}
\arguments{
\item{object}{Function to test.}
}
\description{
Expect tidy format
}
\details{
Test that the format used in a function is tidy (see formatR)
}

