% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expectations.R
\name{expect_tidy_code}
\alias{expect_tidy_code}
\title{Expect tidy format (to be constructed)}
\usage{
expect_tidy_code()
}
\description{
Test that the format used in a function is tidy (see formatR)
}
\keyword{internal}

