% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expectations.R
\name{find_expr}
\alias{find_expr}
\title{Internal function (taken from testthat)}
\usage{
find_expr(name, env = parent.frame())
}
\arguments{
\item{name}{See \code{testthat:::find_expr()}.}

\item{env}{See \code{testthat:::find_expr()}.}
}
\description{
Internal function (taken from testthat)
}

