% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expectations.R
\name{function_code}
\alias{function_code}
\title{Function contain code test}
\usage{
function_code(expected, label = NULL)
}
\arguments{
\item{expected}{Pattern to test for in function code.}

\item{label}{Expectation label used by \code{expect_function_code()}}
}
\description{
Test if function code contains a given text string.
}
\keyword{internal}

