% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_assignment.R
\name{get_file}
\alias{get_file}
\alias{get_file.path_local}
\alias{get_file.path_http}
\alias{get_file.path_github}
\title{Get the file from the path}
\usage{
get_file(path, dest, ...)

\method{get_file}{path_local}(path, dest, ...)

\method{get_file}{path_http}(path, dest, ...)

\method{get_file}{path_github}(path, dest, ...)
}
\arguments{
\item{path}{Path object}

\item{dest}{Destination for the file}

\item{...}{Further arguments to send to \code{httr::GET()}.}
}
\description{
Get/download the file from the path.
}
\keyword{internal}
