% Generated by roxygen2 (4.0.2): do not edit by hand
\name{get_tests}
\alias{get_tests}
\title{Get test files}
\usage{
get_tests(tasks = NULL, force_get_tests = FALSE, verbose = FALSE)
}
\arguments{
\item{tasks}{Which task should be downloaded. Default is "all".}
}
\description{
Get test files
}
\details{
Downloads the test files for the current assignment and save them to
temp directory.
}
\examples{
get_tests(c("uppgift2", "uppgift5"), force_get_tests=TRUE, verbose=TRUE)
 get_tests(force_get_tests=TRUE, verbose=TRUE)
 get_tests(verbose=TRUE)
}

