% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mark_my_assignment.R
\name{get_tests}
\alias{get_tests}
\title{Get test files}
\usage{
get_tests(tasks = NULL, force_get_tests = FALSE)
}
\arguments{
\item{tasks}{Which task should be downloaded. Default is "all".}

\item{force_get_tests}{Force download/get test (ignore cached tests).}
}
\description{
Downloads the test files for the current assignment and save them to 
temp directory.
}

