% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expectations.R
\name{is_self_contained}
\alias{is_self_contained}
\title{Function is self contained test}
\usage{
is_self_contained(expected)
}
\arguments{
\item{expected}{Function to test if it is self contained.}
}
\description{
Tests if a function is self contained (no global variables)
}
\keyword{internal}

