% Generated by roxygen2 (4.0.2): do not edit by hand
\name{mark_my_assignment}
\alias{mark_my_assignment}
\title{Mark assignment in global environment}
\usage{
mark_my_assignment()
}
\arguments{
\item{task}{Which task should be corrected (if more than one).}

\item{file}{Run tests on a R-file. Default is NULL means global environment.}

\item{force_files}{Force download of test files before marking of assignments. Default is FALSE.}
}
\description{
Mark assignment in global environment
}
\details{
Mark assignment in global environment.
}

