% Generated by roxygen2 (4.0.2): do not edit by hand
\name{mark_my_assignment}
\alias{mark_my_assignment}
\title{Mark assignment in global environment}
\usage{
mark_my_assignment(tasks = NULL, mark_file = NULL,
  force_get_tests = FALSE, quiet = FALSE)
}
\arguments{
\item{tasks}{Which task should be corrected (if more than one). Default is all.
To see the different task, see \code{\link{show_tasks}}.}

\item{mark_file}{Run tests on a R-file. Default is NULL means global environment.}

\item{force_get_tests}{Force download of test files before marking of assignments. Default is FALSE.}

\item{quiet}{Should test be run without output?}
}
\description{
Mark assignment in global environment
}
\details{
Mark assignment in global environment.
}
\examples{
\donttest{
assignment_path <-
  paste0(system.file(package = "markmyassignment"), "/extdata/example_assignment.yml.R")
set_assignment(assignment_url)
source(paste0(system.file(package = "markmyassignment"), "/extdata/example_lab_file.R"))
mark_my_assignment()
}
}

