% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mark_my_assignment.R
\name{mark_my_assignment}
\alias{mark_my_assignment}
\title{Mark assignment in global environment}
\usage{
mark_my_assignment(tasks = NULL, mark_file = NULL,
  force_get_tests = FALSE, quiet = FALSE, reporter = NULL)
}
\arguments{
\item{tasks}{Which task should be corrected (if more than one). Default is all. 
To see the different task, see \code{\link{show_tasks}}.}

\item{mark_file}{Argument is deprecated, use mark_my_file instead.}

\item{force_get_tests}{Force download of test files before marking of assignments. Default is FALSE.}

\item{quiet}{Should test be run without output?}

\item{reporter}{to use. Default is the 'summary' or specified in assignment yml file.}
}
\description{
Mark assignment in global environment.
}
\examples{
assignment_path <- 
 paste0(system.file(package = "markmyassignment"), "/extdata/example_assignment01.yml")
set_assignment(assignment_path)
source(paste0(system.file(package = "markmyassignment"), "/extdata/example_lab_file.R"))
mark_my_assignment()

}

