% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mark_my_assignment.R
\name{mark_my_assignment}
\alias{mark_my_assignment}
\title{Mark assignment in global environment}
\usage{
mark_my_assignment(
  tasks = NULL,
  mark_file = NULL,
  force_get_tests = FALSE,
  quiet = FALSE,
  ...
)
}
\arguments{
\item{tasks}{Which task should be corrected (if more than one). Default is all. 
To see the different task, see \code{\link{show_tasks}}.}

\item{mark_file}{Argument is deprecated, use mark_my_file instead.}

\item{force_get_tests}{Argument is deprecated, use \code{set_assignment()} instead.}

\item{quiet}{Should test be run without output?}

\item{...}{further arguments sent to \code{test_dir()}.}
}
\description{
Mark assignment in global environment.
}
\examples{
assignment_path <- 
  file.path(system.file(package = "markmyassignment"), "extdata", "example_assignment01.yml")
set_assignment(assignment_path)
source(file.path(system.file(package = "markmyassignment"), "extdata", "example_lab_file.R"))
mark_my_assignment()

}
