% Generated by roxygen2 (4.0.2): do not edit by hand
\name{mark_my_dir}
\alias{mark_my_dir}
\title{Mark assignments in a directory}
\usage{
mark_my_dir()
}
\arguments{
\item{dir}{Directory with assignments files.}

\item{force_files}{Force download of test files before marking of assignments. Default is FALSE.}
}
\description{
Mark assignments in a directory
}
\details{
Marks assignments in a directory. Stores the results.
}

