% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mark_my_assignment.R
\name{mark_my_dir}
\alias{mark_my_dir}
\title{Mark assignments in a directory}
\usage{
mark_my_dir(directory, tasks = NULL, force_get_tests = FALSE)
}
\arguments{
\item{directory}{Directory with assignments files.}

\item{tasks}{Which task should be corrected (if more than one). 
Default is all. To see the different task, see \code{\link{show_tasks}}.}

\item{force_get_tests}{Force download of test files before marking of assignments. Default is FALSE.}
}
\description{
Marks assignments in a directory. Stores the results.
}

