% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mark_my_dir.R
\name{mark_my_dir}
\alias{mark_my_dir}
\title{Mark assignments in a directory}
\usage{
mark_my_dir(
  directory,
  tasks = NULL,
  assignment_path = NULL,
  force_get_tests = FALSE,
  quiet = FALSE,
  ...
)
}
\arguments{
\item{directory}{Directory with assignments files.}

\item{tasks}{Which task should be corrected (if more than one). Default is all. 
To see the different task, see \code{\link{show_tasks}}.}

\item{assignment_path}{Assignment file to set before marking the assignment (url or local path).}

\item{force_get_tests}{Argument is deprecated, use \code{set_assignment()} instead.}

\item{quiet}{Should test be run without output?}

\item{...}{further arguments sent to \code{test_dir()}.}

\item{mark_file}{Path to the file to mark.}
}
\description{
Marks assignments in a directory. Stores the results.
}
\keyword{internal}
