% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markmyassignment-package.R
\docType{package}
\name{markmyassignment-package}
\alias{markmyassignment}
\alias{markmyassignment-package}
\title{markmyassignment: Tools to work with student assignments}
\description{
Automated testing of R assignments.
}
\details{
Automatic marking of R assignments for students and teachers based on testthat test suites.
}
\author{
\strong{Maintainer}: Mans Magnusson \email{mons.magnusson@gmail.com}

Authors:
\itemize{
  \item Oscar Pettersson
}

}
\keyword{internal}
