% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{package}
\name{markmyassignment}
\alias{markmyassignment}
\alias{markmyassignment-package}
\title{Automatic Marking of R Assignments}
\description{
Automatic marking of R assignments for students and teachers based on testthat test suites.
}

