% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_assignment.R
\name{read_assignment_yml}
\alias{read_assignment_yml}
\title{Load assignment information}
\usage{
read_assignment_yml(path = NULL)
}
\arguments{
\item{path}{\code{path object} from \code{\link{path_type}}}
}
\value{
assignment object
}
\description{
Check if there exist an assignmentfile and then load it.
}

