% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mark_my_assignment.R
\name{run_test_suite}
\alias{run_test_suite}
\title{Run test suite}
\usage{
run_test_suite(caller, tasks = NULL, mark_file = NULL, quiet = FALSE,
  reporter = "summary")
}
\arguments{
\item{caller}{Either "mark_my_assignment" or "mark_my_file"}

\item{tasks}{Which task should be tested}

\item{mark_file}{Run tests on a R-file. Default is NULL means global environment.}

\item{quiet}{Should the output be supressed (only returning test results)}

\item{reporter}{Reporter to use. Standard is student.}
}
\value{
test_suite results
}
\description{
Runs test on the tasks. Always run mandatory tests.
}

