% Generated by roxygen2 (4.0.2): do not edit by hand
\name{set_assignment}
\alias{set_assignment}
\title{Set assignment to mark}
\usage{
set_assignment(path, auth_token = NULL)
}
\arguments{
\item{path}{Path to the yml file}

\item{auth_token}{Authorization token (for github). Not implemented.}
}
\description{
Set assignment to mark
}
\details{
Sets the assignment to mark. Behind the scenes it download the test suite for the assignment.
}
\examples{
set_assignment(path = "https://raw.githubusercontent.com/MansMeg/KursRprgm/master/Labs/Test/mandatory.R")
set_assignment(path = "https://raw.githubusercontent.com/MansMeg/KursRprgm/master/Labs/Test/d1.yml")
}

