% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_assignment.R
\name{set_assignment}
\alias{set_assignment}
\title{Set assignment to mark}
\usage{
set_assignment(path, auth_token = NULL)
}
\arguments{
\item{path}{Path to the yml file}

\item{auth_token}{Authorization token (for github). Not implemented.}
}
\description{
Sets the assignment to mark. Behind the scenes it download the test suite for the assignment.
}
\examples{
assignment_path <- path <- 
  paste0(system.file(package = "markmyassignment"), "/extdata/example_assignment01.yml")
set_assignment(assignment_path)

}

