% Generated by roxygen2 (4.0.2): do not edit by hand
\name{show_tasks}
\alias{show_tasks}
\title{Get the name of the tasks in the assignment.}
\usage{
show_tasks()
}
\description{
Get the name of the tasks in the assignment.
}
\examples{
\donttest{
assignment_path <-
  paste0(system.file(package = "markmyassignment"), "/extdata/example_assignment.yml.R")
set_assignment(assignment_url)
show_tasks()
}
}

