% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_assignment.R
\name{show_tasks}
\alias{show_tasks}
\title{Get the name of the tasks in the assignment.}
\usage{
show_tasks()
}
\description{
Get the name of the tasks in the assignment.
}
\examples{
assignment_path <- 
 paste0(system.file(package = "markmyassignment"), "/extdata/example_assignment01.yml")
 set_assignment(assignment_path)
show_tasks()

}

