% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_assignment.R
\name{temp_folder_check_create}
\alias{temp_folder_check_create}
\title{Check and create folder if missing.}
\usage{
temp_folder_check_create()
}
\description{
Checks if markmyassignment folder exist in R temp directory.
  If not, the folder is created.
}
\keyword{internal}

